Backboard bonus
Implemented the game mechanic whereby every basket, even if it hits the backboard, has a certain chance of enabling a timer that awards bonus points for each basket. Added a canvas on the backboard with animated text displaying the bonus points. Added the ability to change the probability of the event occurring based on the game difficulty