Added CPU challenge
Multiplayer version against the CPU has been implemented. Depending on the game's difficulty, the CPU has a higher chance of scoring a basket (either a perfect shot or a backboard shot). If the player beats the CPU, the coin multiplier varies depending on the difficulty. A new results UI has been implemented that allows you to check whether the player has lost, won, or drawn the match